Clean up everything. And I do mean everything.

+++ src/main/java/vazkii/quark/decoration/block/BlockCandle.java
@@ -130,9 +130,9 @@ public class BlockCandle extends BlockMetaVariants<BlockCandle.Variants> impleme
 	@Override
 	@SideOnly(Side.CLIENT)
 	public void randomDisplayTick(IBlockState stateIn, World worldIn, BlockPos pos, Random rand) {
-		double d0 = (double)pos.getX() + 0.5D;
-		double d1 = (double)pos.getY() + 0.7D;
-		double d2 = (double)pos.getZ() + 0.5D;
+		double d0 = pos.getX() + 0.5D;
+		double d1 = pos.getY() + 0.7D;
+		double d2 = pos.getZ() + 0.5D;
 
 		worldIn.spawnParticle(EnumParticleTypes.SMOKE_NORMAL, d0, d1, d2, 0.0D, 0.0D, 0.0D);
 		worldIn.spawnParticle(EnumParticleTypes.FLAME, d0, d1, d2, 0.0D, 0.0D, 0.0D);
