+++ src/main/java/vazkii/quark/management/client/gui/GuiButtonShulker.java
@@ -28,7 +28,7 @@ public class GuiButtonShulker extends GuiButtonChest {
 			EnumDyeColor dye = ((BlockShulkerBox) shulker.getBlockType()).getColor();
 			int color = ItemDye.DYE_COLORS[dye.getDyeDamage()];
 			Color colorObj = new Color(color).brighter();
-			GlStateManager.color((float) colorObj.getRed() / 255F, (float) colorObj.getGreen() / 255F, (float) colorObj.getBlue() / 255F);
+			GlStateManager.color(colorObj.getRed() / 255F, colorObj.getGreen() / 255F, colorObj.getBlue() / 255F);
 			super.drawIcon(16, 128);
 			GlStateManager.color(1F, 1F, 1F);
 			return;
